install_all script improved

+++ Milestone_3/install_all.sh
@@ -2,31 +2,35 @@
 docker pull solr:9
 docker run -d -p 8983:8983 --name initial_solr -v "$(pwd):/data" solr:9
 
-#### create schemaless core (basic)
+sleep 5
+
 docker exec initial_solr solr create -c media_basic -n data_driven_schema_configs
-
-#### create strict core for intermediate (non-schemaless)
 docker exec initial_solr solr create -c media_intermediate -n basic_configs
-
-#### apply intermediate schema (preferred helper)
-python3 apply_schema.py --core media_intermediate --file intermediate_schema.json
-
 docker exec initial_solr solr create -c semantic_core -n basic_configs
 
-# Add the schema defined at semantic_schema.json
+# Apply schemas
+python apply_schema.py --core media_intermediate --file intermediate_schema.json
 curl.exe -X POST -H "Content-type:application/json" --data-binary "@./semantic_schema.json" http://localhost:8983/solr/semantic_core/schema
 
-# Ensure 'datasets' is installed
-if ! python3 -c "import datasets" 2>/dev/null; then
-	echo "Installing required Python package: datasets"
-	pip install datasets
-fi
+docker cp ./synonyms.txt initial_solr:/var/solr/data/media_intermediate/conf/synonyms.txt
+docker cp ./synonyms.txt initial_solr:/var/solr/data/semantic_core/conf/synonyms.txt
 
-# Run add_poster_urls.py
-python3 scripts/add_poster_urls.py
+#check if file already in data
+if ! [ -f data/movies_series_with_posters.json ]; then
+	# Ensure 'datasets' is installed
+	if ! python3 -c "import datasets" 2>/dev/null; then
+		echo "Installing required Python package: datasets"
+		pip install datasets
+	fi
+	# Run add_poster_urls.py
+	python3 scripts/add_poster_urls.py
+fi
 
 # Index the JSON documents.
 curl.exe -X POST -H "Content-type:application/json" --data-binary "@./data/movies_series_with_posters.json" "http://localhost:8983/solr/semantic_core/update?commit=true"
 curl.exe -X POST -H "Content-type:application/json" --data-binary "@./data/movies_series_with_posters.json" "http://localhost:8983/solr/media_basic/update?commit=true"
 curl.exe -X POST -H "Content-type:application/json" --data-binary "@./data/movies_series_with_posters.json" "http://localhost:8983/solr/media_intermediate/update?commit=true"
 
+curl "http://localhost:8983/solr/admin/cores?action=RELOAD&core=media_intermediate"
+curl "http://localhost:8983/solr/admin/cores?action=RELOAD&core=semantic_core"
+curl "http://localhost:8983/solr/admin/cores?action=RELOAD&core=media_basic"
